refactor: update asset imports to use new directory structure
